Add a data-genus attribute to all components for easier testing
Glamorous generated classnames are impossible to use as selectors in e2e
tests and it becomes impossible to separate two div's where you expect
for example a Box and a Banner. Adding these data attributes makes it possible to discover the root genus components in tests and debugging.

There should be no risk except more DOM created, which, when browser
rendered doesn't really matter that much