Fix broken build:
Catalog depends on UglifyJS 2 via html-webpack-plugin -> html-minifier
and UglifyJS does not support ES6. Recent changes saw us introduce a
not-transpiled package, should-forward-prop, to disable forwarding of some styled-system props
to the underlying HTML Element. This inlines that (very minimal) piece
of code so it gets bundled and transpiled into es5.